+++ sheraf/__init__.py
@@ -4,7 +4,7 @@ from ZODB.DemoStorage import DemoStorage
 from ZODB.FileStorage import FileStorage
 from ZODB.MappingStorage import MappingStorage
 
-from .attributes import Attribute, BaseAttribute, set_read_memoization
+from .attributes import Attribute, set_read_memoization
 from .attributes.blobs import Blob, BlobAttribute
 from .attributes.collections import (
     DictAttribute,